[FIX] functionality of filt products
