+++ example/components/UpsellingPageUI/index.js
@@ -6,14 +6,12 @@ export const UpsellingPageUI = (props) => {
     <div style={{ display: 'flex', flexWrap: 'wrap' }}>
       {
         upsellingProducts.map(product => (
-          <>
-            <div key={product.id} style={{ display: 'inline-flex', flexDirection: 'column', width: '20vw' }}>
-              <img src={product.images} width='100px' />
-              <p>{product.name}</p>
-              <p>${product.price}</p>
-              <button onClick={() => handleAddProductUpselling(product)}>Add</button>
-            </div>
-          </>
+          <div key={product.id} style={{ display: 'inline-flex', flexDirection: 'column', width: '20vw' }}>
+            <img src={product.images} width='100px' />
+            <p>{product.name}</p>
+            <p>${product.price}</p>
+            <button onClick={() => handleAddProductUpselling(product)}>Add</button>
+          </div>
         )
         )
       }
