+++ src/components/UpsellingPage/index.js
@@ -1,11 +1,39 @@
-import React from 'react'
+import React, { useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
 
 import { useOrder } from '../../contexts/OrderContext'
 
 export const UpsellingPage = (props) => {
-  const { UIComponent, onSave, upsellingProducts } = props
+  const { UIComponent, onSave, products } = props
   const [orderState, { addProduct }] = useOrder()
+  const [upsellingProducts, setUpsellingProducts] = useState([])
+
+  useEffect(() => {
+    handleProductsOfCart()
+  }, [products])
+
+  /**
+   * products of the cart
+   */
+  const handleProductsOfCart = () => {
+    const cartProducts = Object.values(orderState.carts).map(cart => {
+      return cart?.products.map(product => {
+        return product
+      })
+    })
+    getUpsellingProducts(cartProducts)
+  }
+
+  /**
+   *
+   * filt products if they are already in the cart
+   * @param {array} cartProducts
+   */
+  const getUpsellingProducts = (cartProducts) => {
+    setUpsellingProducts(products.filter(product => product.upselling && cartProducts.map(cartProduct => {
+      return product.id !== cartProduct.id
+    })))
+  }
 
   /**
    * adding product to the cart from upselling
@@ -13,8 +41,6 @@ export const UpsellingPage = (props) => {
    */
   const handleAddProductUpselling = async (product) => {
     const successful = await addProduct(product)
-    console.log(orderState)
-    console.log(successful)
     if (successful) {
       onSave(product)
     }
@@ -37,5 +63,13 @@ UpsellingPage.propTypes = {
   /**
    * Function to save event
    */
-  onSave: PropTypes.func
+  onSave: PropTypes.func,
+  /**
+   *  Products of the current business is required!
+   */
+  products: PropTypes.array.isRequired
+}
+
+UpsellingPage.defaultProps = {
+  upsellingProducts: []
 }
